Corrects filter on TimeXAxis ticks prop
Showing more than 10 ticks on the x axis can be problematic as dates
can over lap.  Previously we used modulus/3 for just the case where the
type was day and did no correction for month. This allowed over lap in
the case where the date range for the graph was large enough to have
many ticks on the x axis.

We now just calc a modulus value based upon the data’s length / 10 to
ensure that we show the correct number of ticks in all cases.