Slight Doc Update ALSO INCLUDING CI WARNING IGNORE REASONING
Updated docs to use state where a linter was warning state was unused

The ini package vulnerability has an MR up and hopefully will be merged soon. It is deep down in the jest dependencies so without convincing jest to accept a forked copy of ini with the rc package updated, I don't see this going anywhere.

The isometric fetch vulnerability is not imported into glamor at all through the fbjs dependency so IMO it is safe to ignore.